Work around(?) signedness issue by explicit type cast. (I imagine that signal_id wanting and int is not a good idea to change, or at least messy, and likewise bits_one wanting an unsigned int...)

+++ eventloop/correlation.cc
@@ -471,7 +471,7 @@ bool raw_array_correlation_2<Tsingle_correlation,Tsingle,T_correlation,T,n,n1,n2
 #define FCNCALL_MULTI_MEMBER(name) name
 #define FCNCALL_MULTI_ARG(name) name
 #define FCNCALL_CALL_CTRL_WRAP(ctrl,call) ctrl##_active = (call)
-#define FCNCALL_CALL_CTRL_WRAP_ARRAY(ctrl_name,ctrl_non_last_index,ctrl_last_index,call) __active |= ((call) ? (ctrl_name##_active ctrl_non_last_index.set(ctrl_last_index)) , true : false)
+#define FCNCALL_CALL_CTRL_WRAP_ARRAY(ctrl_name,ctrl_non_last_index,ctrl_last_index,call) __active |= ((call) ? (ctrl_name##_active ctrl_non_last_index.set((unsigned int) ctrl_last_index)) , true : false)
 #define FCNCALL_RET_TYPE bool
 #define FCNCALL_INIT     bool __active = false;
 #define FCNCALL_RET      return __active;
